apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "topology-api"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.116.0",
  "sqlalchemy>=2.0.41",
  "uvicorn>=0.35.0",
  "pydantic-settings>=2.10.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.4.1"
]

[tool.setuptools.packages.find]
include = ["app*"]
